Couldn't fix a problem, need help in branch.
The problem I'm having is that I'm trying to get the amenities the residence have although for some reason it doesn't register it. But when I bind the same function I use to get all amenities by the residence to a click button and uses it after everything is rendered, it works! The problem is that some stuff have to render before other things can render but have no clue how to fix it.

Or the problem could be bad code. In the end, I want to switch to another branch so I can develop on something else.