Applied the Footer Compenent Hassan did in App.js
But now it is in a stable version, it should work as intended.
Sorry for having to do a force push but the last version broke so I had to get the last stable version, force the current version to have the last stable version and commit it. I now applied the component to the App rather than in the Home Page.